% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_funzioni_Comuni_gibo.R,
%   R/00_nuove_funzioni.R
\name{aggiungi_codici_post_mutazioni}
\alias{aggiungi_codici_post_mutazioni}
\alias{armonizza}
\title{Crea la lista dei Comuni ticinesi a una determinata data, aggiungendo i codici dei Comuni di una data successiva (sperimentale, in alcuni casi non funziona).}
\usage{
aggiungi_codici_post_mutazioni(listastoricizzata, 
                                      data1 = "2012-12-31", 
                                      data2 = "2013-04-14")

armonizza(listastoricizzata, data1, data2)
}
\arguments{
\item{listastoricizzata}{Un data frame con la lista storicizzata dei Comuni ticinesi}

\item{data1}{Data di riferimento per cui si vuole estrarre la lista dei Comuni ticinesi}

\item{data2}{Data successiva a data1, per aggiungere la colonna bfs_new con i codici dei Comuni che avranno alla data2}
}
\value{
Un data frame con la lista (univoca) dei Comuni ticinesi alla data data1, con l'aggiunta della colonna data2 che indica i codici che i Comuni avranno a quella data (ad esempio dopo fusioni)
}
\description{
Partendo dalla lista storicizzata dei Comuni ticinesi, estrae l'elenco (univoco) dei Comuni alla data1, aggiungendo una colonna con i codici che i Comuni avranno alla data2 (dopo fusioni, ad esempio)

N.B.: \code{armonizza} è una nuova versione di 
\code{aggiungi_codici_post_mutazioni}. È quindi meglio utilizzare
 \code{armonizza} (anche se è da considerare ancora sperimentale). 
 Dovrebbe funzionare con i comuni ticinesi a partire dal 31.12.2004.
}
\examples{
# Carica la lista storicizzata dei Comuni del Ticino
data(comuniTI)

c12_13 <- aggiungi_codici_post_mutazioni(comuniTI, 
                                         data1 = "2012-12-31", 
                                         data2 = "2013-04-14")

dim(c12_13)
str(c12_13)

# Carica la lista storicizzata dei Comuni del Ticino
data(comuniTI)

c12_13 <- armonizza(comuniTI,
                     data1 = "2012-12-31", 
                     data2 = "2013-04-14")

dim(c12_13)
str(c12_13)
}
\references{
\url{http://www.bfs.admin.ch/bfs/portal/it/index/infothek/nomenklaturen/blank/blank/gem_liste/02.html}
}
\seealso{
\code{\link{comuni_stato}}
}
\author{
Gibo
}
