% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comuniTI_package.R
\docType{package}
\name{comuniTI}
\alias{comuniTI}
\title{Crea la lista dei Comuni ticinesi alla data desiderata}
\description{
Partendo dalla lista storicizzata dei Comuni del Ticino, messa a disposizione dall'Ust, questo pacchetto permette di creare la lista dei Comuni ticinesi a una determinata data e anche di collegare due liste che hanno una diversa data di riferimento (questo serve ad esempio a armonizzare dei dati comunali a una lista di Comuni uguale - per creare quindi una serie storica, ad esempio)
}
