% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comuniTI_package.R
\docType{data}
\name{distrettiTI}
\alias{distrettiTI}
\title{Lista storicizzata dei distretti ticinesi}
\format{
Un data frame con 8 righe e 4 colonne
}
\usage{
data(distrettiTI)
}
\description{
Data frame con 8 righe e 4 colonne.
}
\details{
Ogni riga rappresenta un distretto.
Per ogni distretto sono disponibili le colonne seguenti:

\itemize{
  \item districtHistId numero storicizzato del distretto
  \item districtId codice federale del distretto
  \item districtLongName nome lungo del distretto
  \item districtShortName nome corto del distretto
}
}
\keyword{datasets}
