% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_funzioni_Comuni_gibo.R
\name{comuni_stato}
\alias{comuni_stato}
\title{Estrae la lista dei Comuni a una determinata data,
partendo dalla lista storicizzata dei Comuni ticinesi.}
\usage{
comuni_stato(listastoricizzata, data.rif = "2012-12-31")
}
\arguments{
\item{listastoricizzata}{Un data frame con la lista storicizzata dei Comuni ticinesi}

\item{data.rif}{di riferimento dello stato dei Comuni desiderato}
}
\value{
Un data frame con la lista (univoca) dei Comuni a una determinata data
}
\description{
Estrae la lista dei Comuni a una determinata data, partendo dalla lista storicizzata dei Comuni ticinesi (e da una data di riferimento).
}
\examples{
# Carica la lista storicizzata dei Comuni del Ticino
data(comuniTI)
# Elenco dei Comuni del Ticino al 31 dicembre 2012
c12 <- comuni_stato(comuniTI, data.rif="2012-12-31")
str(c12)
}
\references{
\url{http://www.bfs.admin.ch/bfs/portal/it/index/infothek/nomenklaturen/blank/blank/gem_liste/02.html}
}
\seealso{
\code{\link{aggiungi_codici_post_mutazioni}}
}
\author{
Gibo
}
